requirements.txt
aiohttp==2.0.7
amqp==2.1.4
appdirs==1.4.3
async-timeout==1.2.1
billiard==3.5.0.2
celery==4.0.2
chardet==3.0.2
Django==1.11.1
django-celery-beat==1.0.1
django-filter==1.0.2
djangorestframework==3.6.3
kombu==4.0.2
lxml==3.7.3
Markdown==2.6.8
multidict==2.1.4
numpy==1.12.1
packaging==16.8
pandas==0.20.1
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
redis==2.10.5
six==1.10.0
vine==1.1.3
yarl==0.10.2
